Better code - Use of useForm
